making tests green

+++ adapters/repos/db/vector/ssdhelpers/kmeans.go
@@ -100,8 +100,8 @@ func (m *KMeans) Centers() [][]float32 {
 	return m.centers
 }
 
-func (m *KMeans) Encode(point []float32) uint64 {
-	return m.Nearest(point)
+func (m *KMeans) Encode(point []float32) byte {
+	return byte(m.Nearest(point))
 }
 
 func (m *KMeans) Nearest(point []float32) uint64 {
